pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aiida-musconv"
dynamic = ["version"] # read from aiida_musconv/__init__.py
description = "aiida-musconv is a plugin that allows to obtain converged supercell size for an interstitial impurity calculation."
authors = [{name = "Muon  group Parma"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: AiiDA"
]
keywords = ["aiida", "workflows"]
requires-python = ">=3.8"
dependencies = [
    "aiida-core>=2.0,<3",
    "aiida-quantumespresso>=4.2.0",
    "aiida-pseudo",
    "musconv"
]

[project.urls]
Source = "https://github.com/positivemuon/aiida-musconv"

[project.optional-dependencies]
docs = [
    "sphinx",
    "docutils",
    "sphinx-copybutton~=0.3.0",
    "sphinx-book-theme~=0.1.0",
    "sphinx-click~=2.7.1"
]
pre-commit = [
    "pre-commit>=2.21.0"
]
tests = [
    "pgtest~=1.3",
    "pytest~=6.0",
    "pytest-regressions~=2.3"
]

[project.entry-points."aiida.workflows"]
"musconv" = "aiida_musconv.workflows.musconv:MusconvWorkChain"

[tool.flit.module]
name = "aiida_musconv"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests"
]
filterwarnings = [
    "ignore:Creating AiiDA configuration folder.*:UserWarning",
    "ignore::DeprecationWarning:frozendict:",
    "ignore::DeprecationWarning:pkg_resources:",
    "ignore::DeprecationWarning:sqlalchemy_utils:"
]
